// src/utils/types.ts
export type ApiResponse<T> = {
    success: boolean;
    message?: string;
    data?: T;
};

export type Audit = {
    createdDate?: string | number[];
    lastModifiedDate?: string | number[];
};

export type EncounterDto = {
    id?: number;
    patientId: number;
    encounterDate?: string;
    reason?: string;
    status?: "OPEN" | "CLOSED";
    audit?: Audit;
    visitCategory?: string;
    type?: string;

    dateOfService?: string;
    reasonForVisit?: string;
};


export type PatientMedicalHistoryDto = {
    id?: number;
    patientId: number;
    encounterId: number;
    description: string;
    esigned?: boolean;
    signedAt?: string;
    signedBy?: string;
    audit?: {
        createdDate?: string;
        lastModifiedDate?: string;
    };
};